# swagger-jsdoc.yaml
openapi: 3.0.3
info:
  title: Swagger productsstore - OpenAPI 3.0
  description: |-
    This is a sample products Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the products store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of productsstore, then click [here](https://editor.swagger.io/?url=https://productsstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load productsstore OAS 2.0` menu option!_

    Some useful links:
    - [The products Store repository](https://github.com/swagger-api/swagger-productsstore)
    - [The source API definition for the products Store](https://github.com/swagger-api/swagger-productsstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: huanhcph35218@fpt.edu.vn
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:3000/
...
paths:
...
  /products/{id}:
    get:
      tags:
        - products
      summary: Find products by ID
      description: Returns a single products
      operationId: getproductsById
      parameters:
        - name: productsId
          in: path
          description: ID of products to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/products"
            application/xml:
              schema:
                $ref: "#/components/schemas/products"
        "400":
          description: Invalid ID supplied
        "404":
          description: products not found
      security:
        - api_key: []
        - productsstore_auth:
            - write:productss
            - read:productss
    put:
      tags:
        - products
      summary: Cập nhập Products
      description: Cập nhập vào hệ thống
      operationId: updateproducts
      requestBody:
        description: Cập nhập sản phẩm hiện có
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/products"
        required: true
      responses:
        "200":
          description: Cập nhập thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/products"
        "400":
          description: Invalid ID supplied
        "404":
          description: products not found
        "422":
          description: Validation exception
    delete:
      tags:
        - products
      summary: Deletes a products
      description: delete a products
      operationId: deleteproducts
      parameters:
        - name: api_key
          in: header
          description: ""
          required: false
          schema:
            type: string
        - name: productsId
          in: path
          description: products id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Invalid products value
      security:
        - productsstore_auth:
            - write:productss
            - read:productss
...
components:
  schemas:
    products:
      type: object
      properties:
        id:
          type: string
          example: 6607860d63a687b85fbfe28e
        name:
          type: string
          example: Hoàng Cao Huân
        image:
          type: string
          example: huanhcph3528
        price:
          type: string
          example: 1
        description:
          type: string
          example: 1234
        rate:
          type: string
          example: "1"
...